add Game class
const game = new Game;
